+++ src/clefnote.ts
@@ -8,7 +8,7 @@ import { Glyph } from './glyph';
 import { RenderContext } from './types/common';
 import { BoundingBox } from './boundingbox';
 import { ClefType } from './clef';
-import { ModifierContext } from 'modifiercontext';
+import { ModifierContext } from './modifiercontext';
 
 /** ClefNote implements clef annotations in measures. */
 export class ClefNote extends Note {