+++ src/timesignote.ts
@@ -1,7 +1,7 @@
 // [VexFlow](http://vexflow.com) - Copyright (c) Mohit Muthanna 2010.
 // Author Taehoon Moon 2014
 
-import { ModifierContext } from 'modifiercontext';
+import { ModifierContext } from './modifiercontext';
 import { Note } from './note';
 import { TimeSignature, TimeSignatureInfo } from './timesignature';