+++ CHANGELOG.md
@@ -10,9 +10,9 @@ Please help test this beta release and [report issues on GitHub](https://github.
 
 - The tsconfig.json `compilerOptions.target` has been updated to ES6 / ES2015. If you are targeting an older environment such as Internet Explorer, you will need to build directly from source code (and change the target back to ES5).
 - `Stave.setNumLines(n: number)` requires a number. Previously, a string would also work. See: [stave.ts](https://github.com/0xfe/vexflow/blob/master/src/stave.ts) and [#1083](https://github.com/0xfe/vexflow/issues/1083).
-- `StaveNote.addModifier(a: Modifier, b: number): this` throws a RuntimeError if the parameters are reversed.
-- `TickContext.getTickableForVoice(voiceIndex: number): Tickable` was previously named `getTickablesForVoice`. We removed the `s` because the method returns a single Tickable. You will need to update calls to this function if you are upgrading from a build from between April 2020 to August 2021.
-- Changed `Modifier.CATEGORY` to `'modifiers'` instead of `'none'`. `.CATEGORY` is used by VexFlow internally to differentiate objects (i.e., a flexible `instanceof`). See commit: 802bd10a472130d273fc5d6aaaa9707a78121616
+- `Note.addModifier(modifier: Modifier, index?: number): this` now throws a RuntimeError if the parameters are reversed.
+- `TickContext.getTickableForVoice(voiceIndex: number): Tickable` was previously named `getTickablesForVoice(voiceIndex: number): Note`. We removed the `s` because the method returns a single Tickable. You will need to update calls to this function if you are upgrading from a build from between April 2020 to August 2021.
+- Changed `Modifier.CATEGORY` to `'Modifier'` instead of `'none'`. The static property `.CATEGORY` is used by VexFlow internally to differentiate objects.
 
 # 3.0.9 / 2020-04-21